Add function to convert angles from degrees to radians
And use it in a few places.

This creates an interdependency between AngleDegrees and AngleRadians. Since they need to be header-only for performance reasons, we need to define them with interweaved definitions. This way we guarantee that the conversions still get inlined. As discussed with rburema, it's okay to define these two structs in one file then, even though that breaks our code style.

Contributes to issue CURA-7928.
