+++ src/TreeSupport.cpp
@@ -1,4 +1,4 @@
-//Copyright (c) 2020 Ultimaker B.V.
+//Copyright (c) 2021 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #include "Application.h" //To get settings.
@@ -8,7 +8,7 @@
 #include "TreeSupport.h"
 #include "progress/Progress.h"
 #include "settings/EnumSettings.h"
-#include "settings/types/AngleRadians.h" //Creating the correct branch angles.
+#include "settings/types/Angle.h" //Creating the correct branch angles.
 #include "settings/types/Ratio.h"
 #include "utils/IntPoint.h" //To normalize vectors.
 #include "utils/logoutput.h"
