+++ src/infill/SubDivCube.cpp
@@ -1,4 +1,4 @@
-//Copyright (c) 2018 Ultimaker B.V.
+//Copyright (c) 2021 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #include "SubDivCube.h"
@@ -6,7 +6,7 @@
 #include <functional>
 
 #include "../sliceDataStorage.h"
-#include "../settings/types/AngleRadians.h" //For the infill angle.
+#include "../settings/types/Angle.h" //For the infill angle.
 #include "../utils/math.h"
 #include "../utils/polygonUtils.h"
