+++ src/sliceDataStorage.h
@@ -1,4 +1,4 @@
-//Copyright (c) 2018 Ultimaker B.V.
+//Copyright (c) 2021 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #ifndef SLICE_DATA_STORAGE_H
@@ -12,7 +12,7 @@
 #include "SupportInfillPart.h"
 #include "TopSurface.h"
 #include "settings/Settings.h" //For MAX_EXTRUDERS.
-#include "settings/types/AngleDegrees.h" //Infill angles.
+#include "settings/types/Angle.h" //Infill angles.
 #include "settings/types/LayerIndex.h"
 #include "utils/AABB.h"
 #include "utils/AABB3D.h"
